Fix bugs, implement dead code, and improve API consistency
- WPA3/PMF immunity: implement _detectWpa3Only() using SAE AKM and MFPR
  bit detection so skip_immune_networks now functions correctly
- capture_half_handshakes: implement M2-only capture path with active
  attack pivot; add CAP_EAPOL_HALF type; guard toHC22000 against it
- Wire csa_deauth_count into _sendDeauthBurst(); standalone deauth uses
  deauth_burst_count, CSA-paired deauth uses csa_deauth_count
- Add ERR_NOT_ACTIVE guard in setChannel, setTarget, startHopping,
  setActive via _initialized flag set on successful begin()
- EAPOL hardening: validate RSN descriptor type, store and verify M1/M2
  replay counter, handle MGMT_SUB_ASSOC_REQ for earlier client detection
- Fix PCAPNG timestamp overflow: writePcapngPacket takes uint64_t,
  writePcapngRecord uses (uint64_t)millis() * 1000
- Add AttackResultCb fired on PMKID exhaustion and CSA expiry
- Add WigleCsvLogger::appendAp(), derive authStr from enc field,
  extract shared _openWithHeader() helper; fix EnterpriseCsvLogger flush
- Rename FilterCb to TargetFilterCb; remove duplicate PacketCb typedef;
  bump POLITICIAN_MAX_CAPTURED 64->128 to match NvsBssidCache capacity;
  probe_aggr_interval_s uint8->uint16; add isAttacking(), resetStats()
- Remove dead EAPOL defines (KEY_LEN, KEY_IV, KEY_RSC, KEY_ID)
- CI: inline example compilation into build-examples.yml; drop
  compile_examples.sh; add it to .gitignore

+++ src/PoliticianFormat.h
@@ -46,7 +46,7 @@ size_t writePcapngRecord(const HandshakeRecord& rec, uint8_t* buffer, size_t max
  * @param max_len Maximum length of the buffer.
  * @return Number of bytes written, or 0 if buffer is too small.
  */
-size_t writePcapngPacket(const uint8_t* payload, size_t payload_len, int8_t rssi, uint32_t ts_usec, uint8_t* buffer, size_t max_len);
+size_t writePcapngPacket(const uint8_t* payload, size_t payload_len, int8_t rssi, uint64_t ts_usec, uint8_t* buffer, size_t max_len);
 
 } // namespace format
 } // namespace politician
